mimic/requirements.txt
pyaml
colorama
pytest